requirements.txt
asgiref==3.7.2
certifi==2023.5.7
distlib==0.3.6
dj-database-url==2.1.0
Django==3.2.19
filelock==3.12.2
gunicorn==21.2.0
packaging==23.1
Pillow==9.2.0
pipenv==2023.7.3
platformdirs==3.8.0
pyotp==2.9.0
python-decouple==3.1
sqlparse==0.4.4
typing_extensions==4.6.0
tzdata==2023.3
virtualenv==20.23.1
virtualenv-clone==0.5.7
daphne
channels
django-paypal
psycopg2
django-cloudinary-storage
channels-redis==4.1.0
cloudinary==1.34
